chore(build): silence xiph 'opus will be very slow' note + drop redundant cast
Three small cleanups to make `cargo build`, `cargo fmt`, and
`cargo clippy --workspace --all-targets -- -D warnings` all pass cleanly:

1. **Define `OPUS_WILL_BE_SLOW`** in both `harness/build.rs` and
   `harness-deep-plc/build.rs`. Without it, debug builds of the C
   reference's `opus_decoder.c` emit a `#pragma message "...opus will
   be very slow."` informational note, which cc-rs forwards as a
   `cargo:warning=` and pollutes the workspace warning report. The C
   code documents this define as the intended suppression flag.

2. **Drop two `samples as i32` casts** in
   `harness-deep-plc/tests/tier2_snr.rs`. Both `CRefFloatDecoder::decode`
   and `OpusDecoder::decode` already return `Result<i32, i32>`, so the
   cast is a no-op that clippy::unnecessary_cast rightly flags.

3. **Reformat one `assert_eq!`** in `ropus/src/api.rs` to satisfy
   `cargo fmt`.

+++ harness-deep-plc/tests/tier2_snr.rs
@@ -118,7 +118,7 @@ fn decode_with_c_float(packets: &[Vec<u8>], drop_pattern: impl Fn(usize) -> bool
                 .unwrap_or_else(|e| panic!("C decode frame {i} failed: {e}"))
         };
         assert_eq!(
-            samples as i32, FRAME_SIZE,
+            samples, FRAME_SIZE,
             "C decoder frame {i} returned {samples} samples (expected {FRAME_SIZE})"
         );
         out.extend_from_slice(&scratch[..frame_samples]);
@@ -144,7 +144,7 @@ fn decode_with_ropus(packets: &[Vec<u8>], drop_pattern: impl Fn(usize) -> bool) 
                 .unwrap_or_else(|e| panic!("ropus decode frame {i} failed: {e}"))
         };
         assert_eq!(
-            samples as i32, FRAME_SIZE,
+            samples, FRAME_SIZE,
             "ropus decoder frame {i} returned {samples} samples (expected {FRAME_SIZE})"
         );
         out.extend_from_slice(&scratch[..frame_samples]);
